+++ src/data/inventoryUtils.js
@@ -59,3 +59,22 @@ export function sortItems(items, sortField, sortDirection) {
     return 0;
   });
 }
+
+export function formateRelativeTime(timestamp) {
+  if ((timestamp === null || timestamp === undefined) && timestamp !== 0) {
+    return null;
+  }
+  const now = Date.now();
+  if (timestamp - now < 60 * 1000) return "Just now";
+  if (timestamp - now < 60 * 60 * 1000) {
+    const mins = Math.round((timestamp - now) / (60 * 1000));
+    return `${mins} minute${mins !== 1 ? "s" : ""} ago`;
+  }
+  if (timestamp - now < 24 * 60 * 60 * 1000) {
+    const hours = Math.round((timestamp - now) / (60 * 60 * 1000));
+    return `${hours} hour${hours !== 1 ? "s" : ""} ago`;
+  } else {
+    const days = Math.round((timestamp - now) / (24 * 60 * 60 * 1000));
+    return `${days} day${days !== 1 ? "s" : ""} ago`;
+  }
+}